Convert the Classycle plugin to Kotlin
The plugin implementation is quite straightforward but a few things
require an explanation.

In `buildSrc`, the Kotlin DSL can't know what plugins are to be applied.
There's no plugins {} block. This prevents having generated static
Kotlin extensions for Gradle extensions, conventions or configurations.
To make the Kotlin code closer to what you would have in a .gradle.kts
file, the Kotlin extensions on which the Classycle plugin depends, are
declared explicitly, see `accessors.kt`.

The Classycle task makes use of the internal IsolatedAntBuilder that
relies on Groovy method missing, this is handled thanks to the
withGroovyBuilder {} helper provided by the Kotlin DSL.

Signed-off-by: Paul Merlin <paul@gradle.com>