Extract source->generated mapping from annotation processors
Instead of hardcoding the knowledge about what the test processor
does, only hardcode the fact that it exists and extract what it
does dynamically by intercepting the filer. This proves that we
can hook in between javac and the actual processors to get this
information. The next step would be doing this for each processor
on the processor path instead of hardcoding one of them.

Also, the source->generated mapping should be kept in the incremental
compile caches instead of in a file in the build directory. This was
just done to keep the spike simple.